Fix deprecated authenticate issue after  php_mongo 1.2.11

+++ index.php
@@ -19,7 +19,7 @@ error_reporting(E_ALL);
 if (!version_compare(PHP_VERSION, "5.0")) {
 	exit("To make things right, you must install PHP5");
 }
-if (!class_exists("Mongo")) {
+if (!class_exists("Mongo") && !class_exists("MongoClient")) {
 	exit("To make things right, you must install php_mongo module. <a href=\"http://www.php.net/manual/en/mongo.installation.php\" target=\"_blank\">Here for installation documents on PHP.net.</a>");
 }
